// types/Order/OrderStatusEnum.ts
export enum OrderStatusEnum {
  None = 'None',
  Pending = 'Pending',
  Paying = 'Paying',
  PaymentFailed = 'PaymentFailed',
  Paid = 'Paid',
  Processing = 'Processing',
  Shipped = 'Shipped',
  Delivered = 'Delivered',
  Cancelled = 'Cancelled',
  Returned = 'Returned',
  Refunded = 'Refunded'
}